+++ Back/app/api/lote.js
@@ -21,9 +21,11 @@ module.exports = function(app){
     api.salvaLote = (req, res) => {
         const knex = app.conexao.conexaoBDKnex();
         const lote = req.body;
+        const logDAO = new app.infra.LogDAO(knex);
         
         knex('lote').insert(lote)
             .then(resultado => {
+                logDAO.logLote(req.headers['cod_usuario'], 'lote', 'i', null, resultado[0]);
                 knex.destroy();
                 res.status(200).json(resultado[0]);
             })
@@ -39,43 +41,69 @@ module.exports = function(app){
         const { cod_lote } = req.params;
 
         if(cod_lote){
-            const knex = app.conexao.conexaoBDKnex();
             const lote = req.body;
-            
-            knex('lote').where('cod_lote', cod_lote).update(lote)
-                .then(resultado => {
-                    knex.destroy();
-                    res.status(200).json(lote);
-                })
-                .catch(erro => {
+            const connection = app.conexao.conexaoBD();
+            const loteDAO = new app.infra.LoteDAO(connection);
+            loteDAO.listarLoteLog (cod_lote, (erro, resultado) => {
+                if(erro){
                     console.log(erro);
-                    knex.destroy();
                     res.status(500).send(app.api.erroPadrao());
-                });
+                }
+                else{
+                    const knex = app.conexao.conexaoBDKnex();
+                    const logDAO = new app.infra.LogDAO(knex);
+                    let espelho = resultado[0].espelho;
+                    knex('lote').where('cod_lote', cod_lote).update(lote)
+                        .then(resultado => {
+                            logDAO.logLote(req.headers['cod_usuario'], 'lote', 'u', espelho, cod_lote);
+                            knex.destroy();
+                            res.status(200).json(lote);
+                        })
+                        .catch(erro => {
+                            console.log(erro);
+                            knex.destroy();
+                            res.status(500).send(app.api.erroPadrao());
+                        });
+                }
+            });
+            connection.end();
         } else { res.status(400).send(app.api.erroPadrao()); }
     }
     
     //Apaga um Lote.
     api.apagaLote = (req, res) => {
         const { cod_lote } = req.params;
-
+        
         if(cod_lote){
-            const knex = app.conexao.conexaoBDKnex();
-
-            knex('lote').where('cod_lote', cod_lote).delete()
-                .then(resultado => {
-                    knex.destroy();
-                    res.status(200).end();
-                })
-                .catch(erro => {
+            const connection = app.conexao.conexaoBD();
+            const loteDAO = new app.infra.LoteDAO(connection);
+            loteDAO.listarLoteLog(cod_lote, (erro, resultado) => {
+                if(erro){
                     console.log(erro);
-                    knex.destroy();
-                    if(erro.errno == 1451){
-                        res.status(500).send('Este lote não pode ser apagado pois existem outras informações associadas a ele.');
-                    } else {
-                        res.status(500).send(app.api.erroPadrao());
-                    }
-                });
+                    res.status(500).send(app.api.erroPadrao());                    
+                }
+                else{
+                    const knex = app.conexao.conexaoBDKnex();
+                    const logDAO = new app.infra.LogDAO(knex);
+                    let espelho = resultado[0].espelho;
+                    knex('lote').where('cod_lote', cod_lote).delete()
+                        .then(resultado => {
+                            logDAO.logLote(req.headers['cod_usuario'], 'lote', 'd', espelho, cod_lote);
+                            knex.destroy();
+                            res.status(200).end();
+                        })
+                        .catch(erro => {
+                            console.log(erro);
+                            knex.destroy();
+                            if(erro.errno == 1451){
+                                res.status(500).send('Este lote não pode ser apagado pois existem outras informações associadas a ele.');
+                            } else {
+                                res.status(500).send(app.api.erroPadrao());
+                            }
+                        });
+                }
+            });
+            connection.end();
         } else { res.status(400).send(app.api.erroPadrao()); }
     }
